slackbot/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==2.1.0; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
flask==2.1.3
idna==2.8
importlib-metadata==4.12.0; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
parse==1.19.0
pip==22.2; python_version >= '3.7'
python-dotenv==0.20.0
requests==2.28.1; python_version >= '3.7' and python_version < '4'
setuptools==63.2.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
slack-bolt==1.14.2
slack-bot==0.0.7
slack-sdk==3.18.1
slackclient==1.3.1
urllib3==1.24.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3' and python_version < '4'
websocket-client==0.47.0
werkzeug==2.2.0; python_version >= '3.7'
zipp==3.8.1; python_version >= '3.7'
